Encapsulating test exectution

+++ WSN_UnitTests/Test_SensorNetwork.cpp
@@ -3,13 +3,12 @@
 #include "NetworkLayer.h"
 #include "NetworkNode.h"
 
-
 void Test_SensorNetwork::test_addLayer()
 {
     SensorNetwork network;
-    QCOMPARE(network.getNumberOfLayers(), 0);
+    QCOMPARE(network.getNumberOfLayers(), static_cast<quint16>(0));
     network.addLayer();
-    QCOMPARE(network.getNumberOfLayers(), 1);
+    QCOMPARE(network.getNumberOfLayers(), static_cast<quint16>(1));
 }
 
 void Test_SensorNetwork::test_createNode()
